createproject both html and php
I made it into a form so that it collects the data and sends the title and description into the database. This hasn't been properly tested because I can't connect to a localhost on this computer for some reason and I might need someone else to test it out before I can do anything
Also note that the id for all of the entries will be "4" because I didn't know if it needed an id to be placed into the database.